+++ packages/storage/src/tabular/SharedInMemoryTabularRepository.ts
@@ -4,13 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import {
-  createServiceToken,
-  DataPortSchemaObject,
-  ExcludeProps,
-  FromSchema,
-  IncludeProps,
-} from "@podley/util";
+import { createServiceToken, DataPortSchemaObject, FromSchema } from "@podley/util";
 import { ITabularRepository } from "./ITabularRepository";
 import { InMemoryTabularRepository } from "./InMemoryTabularRepository";
 import { TabularRepository } from "./TabularRepository";
@@ -43,17 +37,17 @@ export class SharedInMemoryTabularRepository<
   Schema extends DataPortSchemaObject,
   PrimaryKeyNames extends ReadonlyArray<keyof Schema["properties"]>,
   // computed types
-  PrimaryKey = FromSchema<IncludeProps<Schema, PrimaryKeyNames>>,
   Entity = FromSchema<Schema>,
-  Value = FromSchema<ExcludeProps<Schema, PrimaryKeyNames>>,
-> extends TabularRepository<Schema, PrimaryKeyNames, PrimaryKey, Entity, Value> {
+  PrimaryKey = Pick<Entity, PrimaryKeyNames[number] & keyof Entity>,
+  Value = Omit<Entity, PrimaryKeyNames[number] & keyof Entity>,
+> extends TabularRepository<Schema, PrimaryKeyNames, Entity, PrimaryKey, Value> {
   private channel: BroadcastChannel | null = null;
   private channelName: string;
   private inMemoryRepo: InMemoryTabularRepository<
     Schema,
     PrimaryKeyNames,
-    PrimaryKey,
     Entity,
+    PrimaryKey,
     Value
   >;
   private isInitialized = false;
@@ -78,8 +72,8 @@ export class SharedInMemoryTabularRepository<
     this.inMemoryRepo = new InMemoryTabularRepository<
       Schema,
       PrimaryKeyNames,
-      PrimaryKey,
       Entity,
+      PrimaryKey,
       Value
     >(schema, primaryKeyNames, indexes);